+++ aa-display/src/main/java/io/github/nitsuya/aa/display/xposed/CoreManagerService.kt
@@ -372,3 +372,14 @@ class CoreManagerService private constructor() : ICoreManager.Stub() {
 }
 
 
+
+
+
+
+
+
+
+
+
+
+